+++ ios/Plugin/SecureStoragePlugin.m
@@ -11,4 +11,6 @@ CAP_PLUGIN(SecureStoragePlugin, "SecureStoragePlugin",
            CAP_PLUGIN_METHOD(remove, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(clear, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(getPlatform, CAPPluginReturnPromise);
+           CAP_PLUGIN_METHOD(setStorageAdapter, CAPPluginReturnPromise);
+           CAP_PLUGIN_METHOD(getStorageAdapter, CAPPluginReturnPromise);
 )