+++ src/web.ts
@@ -1,12 +1,45 @@
 import { WebPlugin } from '@capacitor/core';
-import { SetDataOptions } from '.';
-import { Accessibility, SecureStoragePlugin } from './definitions';
+
+import type { SetDataOptions } from '.';
+import type { Accessibility} from './definitions';
+import { SecureStoragePlugin, WebStorageAdapter } from './definitions';
+import { sessionCookie } from './sessionCookie'
 
 export class SecureStoragePluginWeb extends WebPlugin implements SecureStoragePlugin {
   PREFIX = 'cap_sec_';
+  STORAGE_ADAPTER_KEY = 'cap_sec_adapter'
+
+  private getCurrentStorageAdapter() {
+    const adapter = localStorage.getItem(this.STORAGE_ADAPTER_KEY)
+
+    switch(adapter) {
+      case WebStorageAdapter.SessionStorage:
+        return sessionStorage
+      case WebStorageAdapter.SessionCookie:
+        return sessionCookie
+      default:
+        return localStorage
+    }
+  }
+
+  async setStorageAdapter(options: { web: {adapter: WebStorageAdapter} }): Promise<{ value: boolean }> {
+    localStorage.setItem(this.STORAGE_ADAPTER_KEY, options.web.adapter)
+    return {
+      value: true
+    }
+  }
+
+  async getStorageAdapter(): Promise<{ adapter: WebStorageAdapter }> {
+    const adapter = (localStorage.getItem(this.STORAGE_ADAPTER_KEY) || WebStorageAdapter.LocalStorage) as WebStorageAdapter
+
+    return {
+      adapter
+    }
+  }
 
   get(options: { key: string }): Promise<{ value: string }> {
-    const item = localStorage.getItem(this.addPrefix(options.key));
+    const storage = this.getCurrentStorageAdapter()
+    const item = storage.getItem(this.addPrefix(options.key));
     if (item !== null) {
       return Promise.resolve({
         value: atob(item),
@@ -22,23 +55,31 @@ export class SecureStoragePluginWeb extends WebPlugin implements SecureStoragePl
   }
 
   set(options: SetDataOptions): Promise<{ value: boolean }> {
-    localStorage.setItem(this.addPrefix(options.key), btoa(options.value));
+    const storage = this.getCurrentStorageAdapter()
+    storage.setItem(this.addPrefix(options.key), btoa(options.value));
     return Promise.resolve({ value: true });
   }
   remove(options: { key: string }): Promise<{ value: boolean }> {
-    localStorage.removeItem(this.addPrefix(options.key));
+    const storage = this.getCurrentStorageAdapter()
+    storage.removeItem(this.addPrefix(options.key));
     return Promise.resolve({ value: true });
   }
   clear(): Promise<{ value: boolean }> {
-    for (const key in localStorage) {
+    const storage = this.getCurrentStorageAdapter()
+    for (const key in storage) {
       if (key.indexOf(this.PREFIX) === 0) {
-        localStorage.removeItem(key);
+        storage.removeItem(key);
       }
     }
+
+    // remove eventually set storageAdapter key
+    localStorage.removeItem(this.STORAGE_ADAPTER_KEY)
+
     return Promise.resolve({ value: true });
   }
   keys(): Promise<{ value: string[] }> {
-    const keys = Object.keys(localStorage).filter((k) => k.indexOf(this.PREFIX) === 0);
+    const storage = this.getCurrentStorageAdapter()
+    const keys = Object.keys(storage).filter((k) => k.indexOf(this.PREFIX) === 0);
     return Promise.resolve({ value: keys });
   }